refine

+++ paddle/phi/kernels/onednn/shape_kernel.cc
@@ -36,7 +36,7 @@ void ShapeKernel(const Context& dev_ctx,
   }
 
   out->Resize({x_dims.size()});
-  auto out_data = dev_ctx.template Alloc<int32_t>(out);
+  auto out_data = dev_ctx.template Alloc<int64_t>(out);
   for (int i = 0; i < x_dims.size(); ++i) {
     out_data[i] = x_dims[i];
   }
